Hashpipe.jl's Documentation

Hashpipe.Hashpipe — Module

Hashpipe Module

C Hashpipe functions for the High-Availablity Shared Pipeline Engine ported for Julia usability with pure-Julian extensions. Written by Max Hawkins Hashpipe C code written by Dave MacMahon: https://github.com/david-macmahon/hashpipe

Hashpipe.HASHPIPE_ERR_GEN — Constant

HASHPIPE_ERR_GEN

Status code signifying an unknown error?

Hashpipe.HASHPIPE_ERR_KEY — Constant

HASHPIPE_ERR_KEY

Status code signifying requested key doesn't exist.

Hashpipe.HASHPIPE_ERR_PACKET — Constant

HASHPIPE_ERR_PACKET

Status code signifying unexpected packet size.

Hashpipe.HASHPIPE_ERR_PARAM — Constant

HASHPIPE_ERR_PARAM

Status code signifying parameter out of valid range.

Hashpipe.HASHPIPE_ERR_SYS — Constant

HASHPIPE_ERR_SYS

Status code signifying a failed system call.

Hashpipe.HASHPIPE_OK — Constant

HASHPIPE_OK

Status code signifying successful operation.

Hashpipe.HASHPIPE_TIMEOUT — Constant

HASHPIPE_TIMEOUT

Status code signifying call timeout.

Hashpipe.MAX_HASHPIPE_THREADS — Constant

MAX_HASHPIPE_THREADS

Maximum number of threads that be defined by plugins.

Hashpipe.STATUS_RECORD_SIZE — Constant

const global STATUS_RECORD_SIZE = 80

The size of a status record - 80 characters.

Hashpipe.STATUS_TOTAL_SIZE — Constant

const global STATUS_TOTAL_SIZE = 184320

The maximum size of a HASHPIPE status buffer - 184,320 bytes.

Hashpipe.THREAD_ERROR — Constant

THREAD_ERROR

Error status from run.

Hashpipe.THREAD_OK — Constant

THREAD_OK

OK status from run.

Hashpipe.HashpipeThread — Type

HashpipeThread

Abstract type describing a Hashpipe thread.

name - A string containing the thread's name. Used to match command line
        thread spcifiers to thread metadata so that the pipeline can be 
        constructed as specified on the command line.
skey - A string containing the thread's status buffer "status" key.
        Typically 8 characters or less, uppercase, and ends with "STAT".
        If it is non-NULL and non-empty, HASHPIPE will automatically 
        store/update this key in the status buffer with the thread's 
        status at initialization ("init") and exit ("exit").
init - A pointer to the thread's initialization function.
run  - A pointer to the thread's run function.
ibuf - A structure describing the thread's input data buffer (if any).
obuf - A structure describing the thread's output data buffer (if any).

Description of Hashpipe threads

A hashpipe_thread structure encapsulates metadata and functionality for one or more threads that can be used in a processing pipeline. The hashpipe executable dynamically assembles a pipeline at runtime consisting of multiple hashpipe threads.

Hashpipe threads must register themselves with the hashpipe executable via a call to registerhashpipethread(). This is typically performed from a static C function with the constructor attribute in the hashpipe thread's source file.

Hashpipe threads are identified by their name. The hashpipe executable finds (registered) hashpipe threads by their name. A hashpipe thread can be input-only, output-only, or both input and output. An input thread has an associated output data buffer into which it writes data. An output thread has an associated input data buffer from which it reads data. An input/output thread has both.

Input-only threads source data into the pipeline. They do not get their input data from a shared memory ring buffer. They get their data from external sources (e.g. files or the network) or generate it internally (e.g. for test vectors). Input-only threads have an output data buffer, but no input data buffer (their input does not come from a shared memory ring buffer).

Output-only threads sink data from the pipeline. Thy do not put their output data into a shared memory ring buffer. They send their data to external sinks (e.g. files or the network) of consume it internally (e.g. comparing against expected output). Output-only threads have an input data buffer, but no output data buffer (their output data does not go the a shared memory ring buffer).

Input/output threads get their input data from one shared memory region (their input data buffer), process it, and store the output data in another shared memory region (their output data buffer).

Hashpipe.databuf_desc_t — Type

struct databuf_desc_t

Contain the pointer to the databuf create function.

Hashpipe.databuf_t — Type

databuf_t <: HashpipeDatabuf

Contain data pertaining to a Hashpipe databuffer.

Hashpipe.status_t — Type

Hashpipe Status struct

Data representing a Hashpipe status buffer.

Need to create empty status struct before trying to attaching to existing status buffer. Example: ''' instanceid = 0 status = statust(0,0,0,0) statusattach(instanceid, Ref(status)) '''

Hashpipe.thread_args_t — Type

thread_args_t

Implementation data for a Hashpipe thread.

Hashpipe.thread_desc_t — Type

thread_desc_t

Metadata describing a Hashpipe thread (usually defined by a plugin per thread).

The thread_desc structure is used to store metadata describing a hashpipe thread. Typically a hashpipe plugin will define one of these hashpipe thread descriptors per hashpipe thread.

Base.Multimedia.display — Method

Base.display(p::Ptr{databuf_t})

Display Hashpipe databuffer from pointer nicely in REPL.

Base.Multimedia.display — Method

Base.display(r::Ref{status_t})

Display Hashpipe status from reference nicely in REPL.

Base.Multimedia.display — Method

Base.display(d::Hashpipe.databuf_t)

Display a Hashpipe databuffer nicely in REPL.

Base.Multimedia.display — Method

Base.display(s::status_t)

Display a Hashpipe status nicely in REPL.

Hashpipe.check_databuf — Function

check_databuf(instance_id=0, db_id=1)

Display databuf information of a given databuf of a Hashpipe instance.

Hashpipe.databuf_attach — Method

databuf_attach(instance_id::Int, db_id::Int)

Attach a databuf to a Hashpipe instance.

Hashpipe.databuf_block_status — Method

databuf_block_status(p_databuf::Ptr{databuf_t}, block_id::Int)

Return the status of the selected data block of the given databuffer.

Hashpipe.databuf_clear — Method

databuf_clear(p_databuf::Ptr{databuf_t})

Clear the data in a databuf.

Hashpipe.databuf_create — Method

databuf_create(instance_id::Int, db_id::Int, header_size::Int, block_size::Int, n_block::Int)

Create a databuffer with given parameters.

Hashpipe.databuf_data — Method

databuf_data(p_databuf::Ptr{databuf_t}, block_id::Int)

Return the pointer to the associated databuffer's block's data.

Hashpipe.databuf_detach — Method

databuf_detach(instance_id::Int, db_id::Int)

Detach a databuf from a Hashpipe instance.

Hashpipe.databuf_set_filled — Method

databuf_set_filled(p_databuf::Ptr{databuf_t}, block_id::Int)

Set the given block of data as filled. Return an error if the block is already filled.

See also: databuf_wait_filled, databuf_wait_free, databuf_set_free

Hashpipe.databuf_set_free — Method

databuf_set_free(p_databuf::Ptr{databuf_t}, block_id::Int)

Set the given block of data as free. Return an error if the block is already free.

See also: databuf_wait_filled, databuf_wait_free, databuf_set_filled

Hashpipe.databuf_total_mask — Method

databuf_total_mask(p_databuf::Ptr{databuf_t})

Hashpipe.databuf_total_status — Method

databuf_total_status(p_databuf::Ptr{databuf_t})

Return the total lock status for the given databuffer.

Hashpipe.databuf_wait_filled — Method

databuf_wait_filled(p_databuf::Ptr{databuf_t}, block_id::Int)

Wait for the given block of data to be filled.

See also: databuf_wait_free, databuf_set_filled, databuf_set_free

Hashpipe.databuf_wait_free — Method

databuf_wait_free(p_databuf::Ptr{databuf_t}, block_id::Int)

Wait for the given block of data to be freed.

See also: databuf_wait_filled, databuf_set_filled, databuf_set_free

Hashpipe.find_hashpipe_thread — Method

find_hashpipe_thread(name)

Find a Hashpipe thread with the given name.

Generally used only by the hashpipe executable. Returns a pointer to its hashpipethreaddesc_t structure or NULL if a test with the given name is not found.

Names are case-sensitive.

Hashpipe.get_cpu_affinity — Method

get_cpu_affinity()

Hashpipe.register_hashpipe_thread — Method

register_hashpipe_thread(ptm)

Register a Hashpipe thread with the pipeline executable.

Hashpipe.run_threads — Method

run_threads()

Function that threads use to determine whether to keep running.

Hashpipe.status_attach — Method

status_attach(instance_id::Int, p_status::Ref{status_t})

Populate the Hashpipe status pointed to by p_status with the status values of the Hashpipe instance given (created if doesn't already exist).

Attach/create lock semaphore as well. Return nonzero on error.

Hashpipe.status_buf_lock_unlock — Method

status_buf_lock_unlock(f::Function, r_status::Ref{status_t})

Safely lock and unlock a shared status buffer for updating its values. This must be done so that the status buffer values aren't changed by multiple processes at the same time.

Example:

Hashpipe.status_chkinit — Method

status_chkinit(p_status::Ref{status_t})

Check the status buffer for appropriate formatting (existence of "END"). If not found, zero it out and add END.

Hashpipe.status_clear — Method

status_clear(p_status::Ref{status_t})

Clear the status values of the status buffer shared memory.

Hashpipe.status_exists — Method

status_exists(instance_id)

Check whether or not the Hashpipe status buffer exists for the given Hashpipe instance.

Returns non-zero if the status buffer for instance already exists.

Hashpipe.status_lock — Method

status_lock(p_status::Ref{status_t})

Lock the status pointed to by p_status (probably for updating status values).

If locked, will sleep while waiting for the buffer to become unlocked.

Hashpipe.status_unlock — Method

status_unlock(p_status::Ref{status_t})

Unlock the status pointed to by p_status (probably after updating status values).

If unlocked, will sleep while waiting for the buffer to become locked.

Hashpipe.Hashpipe

Hashpipe.HASHPIPE_ERR_GEN

Hashpipe.HASHPIPE_ERR_KEY

Hashpipe.HASHPIPE_ERR_PACKET

Hashpipe.HASHPIPE_ERR_PARAM

Hashpipe.HASHPIPE_ERR_SYS

Hashpipe.HASHPIPE_OK

Hashpipe.HASHPIPE_TIMEOUT

Hashpipe.MAX_HASHPIPE_THREADS

Hashpipe.STATUS_RECORD_SIZE

Hashpipe.STATUS_TOTAL_SIZE

Hashpipe.THREAD_ERROR

Hashpipe.THREAD_OK

Hashpipe.HashpipeThread

Hashpipe.databuf_desc_t

Hashpipe.databuf_t

Hashpipe.status_t

Hashpipe.thread_args_t

Hashpipe.thread_desc_t

Base.Multimedia.display

Base.Multimedia.display

Base.Multimedia.display

Base.Multimedia.display

Hashpipe.check_databuf

Hashpipe.databuf_attach

Hashpipe.databuf_block_status

Hashpipe.databuf_clear

Hashpipe.databuf_create

Hashpipe.databuf_data

Hashpipe.databuf_detach

Hashpipe.databuf_set_filled

Hashpipe.databuf_set_free

Hashpipe.databuf_total_mask

Hashpipe.databuf_total_status

Hashpipe.databuf_wait_filled

Hashpipe.databuf_wait_free

Hashpipe.find_hashpipe_thread

Hashpipe.get_cpu_affinity

Hashpipe.register_hashpipe_thread

Hashpipe.run_threads

Hashpipe.status_attach

Hashpipe.status_buf_lock_unlock

Hashpipe.status_chkinit

Hashpipe.status_clear

Hashpipe.status_exists

Hashpipe.status_lock

Hashpipe.status_unlock